requirements.txt
django
djangorestframework
pillow
django-extra-fields